requirements.txt
fastapi
uvicorn
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pydantic
pandas
pyarrow
fastavro